dont include volumes , as we dont double count them even with flag
